insercao funcao verificar erros e quando cancela

+++ lib/firebase/bancoDados.dart
@@ -11,6 +11,30 @@ class BancoDadosFirebase {
   String imageUrl = '';
   TaskSnapshot? photeRemov;
 
+  Future<Map<String, dynamic>> getNomeFoto(String documentId) async {
+    try {
+      DocumentSnapshot document = await FirebaseFirestore.instance
+          .collection('users')
+          .doc(documentId)
+          .get();
+
+      // O documento existe, você pode acessar os dados dentro dele
+      Map<String, dynamic> data = document.data() as Map<String, dynamic>;
+
+      //String nome = data['nome'];
+      //String fotoUrl = data['foto'];
+      return data;
+
+      // Faça o que você precisa com o nome e a foto aqui
+    } catch (e) {
+      return {
+        'nome': 'genmarc',
+        'foto':
+            'https://img.freepik.com/vetores-gratis/desenho-de-carrinho-e-construcao-de-loja_138676-2085.jpg?w=2000',
+      };
+    }
+  }
+
   // Função para verificar a existência de um documento
   Future<bool> isDocumentExist(String documentId) async {
     final DocumentSnapshot document = await FirebaseFirestore.instance
@@ -33,12 +57,44 @@ class BancoDadosFirebase {
       // Crie um mapa com os dados que você deseja adicionar ao documento
       Map<String, dynamic> data = {
         'nome': nome,
-        'foto': imageUrl,
+        'foto': pathImage == null
+            ? 'https://img.freepik.com/vetores-gratis/desenho-de-carrinho-e-construcao-de-loja_138676-2085.jpg?w=2000'
+            : imageUrl,
         // Adicione outros campos e valores conforme necessário
       };
 
       // Use o método `set()` para criar um novo documento na coleção ou substituir um documento existente com os dados especificados
       await documentReference.set(data);
+
+      // Subcoleção 1
+      CollectionReference subCollection1 =
+          documentReference.collection('fiado');
+      Map<String, dynamic> dataSubCollection1 = {
+        'campo1': 'valor1',
+        'campo2': 'valor2',
+        // Adicione outros campos e valores conforme necessário para a primeira subcoleção
+      };
+      await subCollection1.add(dataSubCollection1);
+
+// Subcoleção 2
+      CollectionReference subCollection2 =
+          documentReference.collection('empresa');
+      Map<String, dynamic> dataSubCollection2 = {
+        'campo3': 'valor3',
+        'campo4': 'valor4',
+        // Adicione outros campos e valores conforme necessário para a segunda subcoleção
+      };
+      await subCollection2.add(dataSubCollection2);
+
+// Subcoleção 3
+      CollectionReference subCollection3 =
+          documentReference.collection('bancodados');
+      Map<String, dynamic> dataSubCollection3 = {
+        'campo5': 'valor5',
+        'campo6': 'valor6',
+        // Adicione outros campos e valores conforme necessário para a terceira subcoleção
+      };
+      await subCollection3.add(dataSubCollection3);
 
       print('Documento criado ou atualizado com sucesso.');
     } catch (e) {
@@ -59,7 +115,6 @@ class BancoDadosFirebase {
     } else {
       // O usuário cancelou a seleção da imagem
       pathImage = null;
-      print('não pegou');
     }
     await uploadImageToStorageAndFirestore();
   }
@@ -79,13 +134,12 @@ class BancoDadosFirebase {
       } catch (e) {
         print('O erro foi $e');
       }
-    } else {
-      imageUrl =
-          'https://img.freepik.com/vetores-gratis/desenho-de-carrinho-e-construcao-de-loja_138676-2085.jpg?w=2000';
     }
   }
 
   Future<void> deletePhoto() async {
-    await photeRemov!.ref.delete();
+    if (imageUrl != '') {
+      await photeRemov!.ref.delete();
+    }
   }
 }
